Remove all path additions
This commit updates Issue #21 by removing all usage of `addpath()`, so
that we no longer need to reason inside DynaSim whether it has been
installed correctly. The deleted code is unnecessary in the event that
someone has followed the installation guidelines (which have been
updated on README.md) and installed a `startup.m` file.

If someone has installed DynaSim via the `startup.m` method, and their
mode/mechanism files are anywhere in the `models` folder, there is no
longer any need to add the DynaSim path to the MATLAB path; this means
`SimulateModel.m`, `PlotData.m`, and `CreateBatch.m` no longer need to
`addpath()`. Additionally, `SimulateModel.m` and `PlotData.m` no longer
need to be in any specific directory.

Changes to `LocateModelFiles.m` are forthcoming.

Changes to be committed:
	modified:   README.md
	modified:   functions/CreateBatch.m
	renamed:    PlotData.m -> functions/PlotData.m
	renamed:    SimulateModel.m -> functions/SimulateModel.m

 Untracked files:
	models/z048-github-matlab-ching2010-tcre-dynasim-mechanisms/