Remove most path-handling code
This commit updates Issue #21 by removing most of the remaining custom
path-handling code in `LocateModelFiles.m`, and opting to use the
built-in `exist()` and `which()` more liberally. This also removes usage
of `dynasim_path` except for where the latter is simply used as metadata
in `CheckHostPaths.m` Lines 25-34 and `CreateBatch.m` Lines
115-119/134-137.

Previously, there were minor bugs in the use of `search_paths` in
`LocateModelFiles.m`, in addition to it being effectively redundant in
comparison to MATLAB's "built-in" searchpath using `exist()` and
`which()`.

`search_paths`'s path order goes to the

1. local directory,
2. then the DynaSim directory and its subdirectories (but excluding very
long folder names due to a bug),
3. then an empty string (due to a bug in the `regexp()` call, which is
important),
4. then the MATLAB path.

In (mild) contrast, the MATLAB searchpath used by `exist()` and
`which()` goes

1. local directory (I've tested this)
2. if the DynaSim directory is installed correctly at the top of the
MATLAB path e.g. at the top of `startup.m`, then the DynaSim directory
and all its recursive subfolders
3. then the rest of the MATLAB path.

`search_path`'s resulting path order is only different from the MATLAB
searchpath on a system with DynaSim installed via `startup.m` in the
following ways:

1. The DynaSim directory is artificially forced to, effectively, be at
the top of the user-set MATLAB path.
    - However, this is identical to if DynaSim is the first toolbox
      loaded on the MATLAB path, e.g. by installing the `startup.m` file
      correctly, and putting DynaSim first in the file. In other words,
      this is identical to if we force people to load the DynaSim
      toolbox first, before any other toolboxes, in the install.
2. Due to an unknown string bug, all folders and subfolders in the
DynaSim folder are added to `search_paths` EXCEPT for folders that have
very long names, at least on Linux (tested on the cluster and a CompNet
Ubuntu 16.04 LTS desktop). This has not been caught before due to
difference number 3 below, and I happen to have a folder in `models`
that has a very long name. This means that `search_paths` does NOT
correctly gather folders and recursive subfolders in the DynaSim
directory.
    - In contrast, when DynaSim is installed the correct way with
      `startup.m`, these very-long-name folders ARE indexed by MATLAB
      and the files in them CAN be found by `exist()` and `which()` --
      which are only used as a last resort in this code here in
      `LocateModelFiles.m`, but are the only reason my models have been
      used correctly this whole time!
3. Due to the final post-colon split by the `regexp()` call, the last
folder location returned by it is simply an empty string. This is
important because in the for loops below, this indicates a mechanism
file without a supervising folder. This means that even if
`s==num_paths` is false (which it will be because there are more
search_paths to go from the subsequent MATLAB path directories), `mech`
will just be the extension-less mechanism filename alone. Then, because
`exist()` uses the MATLAB searchpath and NOT the artificial
`search_paths` here, if it is anywhere in the (almost identical) MATLAB
searchpath order, it WILL find it, but because its `fileparts()` will be
empty, the built-in `which()` will be used to find the file, which also
uses the MATLAB searchpath and NOT our custom `search_path`. Therefore,
most of `search_paths`, i.e. the post-DynaSim and post-empty string
"MATLAB path" portion of `search_paths`, will NEVER even be used --
unless `exist()/which()` are unable to find any mechanism by that name
under any valid extension. In this case, finding the mechanism fails as
it is not in ANY path, and it fails silently (!!!) by simply passing an
empty string to files.
    - Furthermore, this silent error is passed on to files, and e.g.
      batch simulations are created and sent to the batch system as if
      the model is well-defined, which it is not, since one of the
      mechanisms is missing. We need to add error handling here
      eventually.

Therefore, due to the EXTREMELY HIGH similarity of `search_paths`'s
paths order and the native MATLAB searchpath on a system with a
**properly-installed DynaSim** (except for the bugs mentioned above, which
cause `LocateModelFiles.m` to simply fall back to using the built-in
`exist()` and `which()` functions), I've written `search_paths` out so
the bottom loop relies directly on `exist()`, `which()`, and the native
MATLAB searchpath, instead of our own custom one.

There is no mechanism file that would be found the old `search_paths`
way and not this current way, assuming DynaSim is installed using
`startup.m`, and the order by which MATLAB takes precedence in finding
the mechanism is the same. THUS, we can ignore the relative hierarchy of
folders in the DynaSim code when searching for mechanism files, and we
can now just say "DynaSim searches the MATLAB path as returned by
`path()` for DynaSim mechanism files".

We can leverage MATLAB's path-ing itself to be necessary and sufficient
for communicating how we locate mechanism files! And if, say, you want a
certainly identically-named mechanism in one model to take precedence
over another one, you can just add the folder containing that mechanism
first in the `startup.m` file. I should probably write this down
somewhere in the help docs.

Changes to be committed:
	modified:   functions/LocateModelFiles.m