Add custom m2html dependency and first round docs
This adds a modified m2html dependency, where each use of the MATLAB
2016b-deprecated `nargchk` has been replaced with `narginchk` etc. This
change also speeds up the documentation generation process by a factor
of ten, and thus all the documentation builds in about a minute on a
modern CPU. I think the previous speed issue was due to warnings.

This updates #18.