iss #138 Wrote mexpath option for SimStudy"
This allows the user to create a folder (e.g. ./mexes) to serve as an archive for all of their mex files. DynaSim
will search this folder for compatible mex files when running in compile mode.

The default location of this folder is ./mexes, but this can be changd via input argument to
SimStudy, or disabled by setting it to [].

This closes #138