Add new plot/calc function for traditional PAC
This adds a new plotting type 'comodulogramsBH' and a new analysis function,
'functions/internal/dsCalcComodulogramsBH.m'. This new functionality adds an
option to compute phase-amplitude coupling comodulograms based on the
traditional Modulation Index method of filtering the signals for their
respective frequencies using a Butterwoth filter, and then using the Hilbert
transform to obtain their instantaneous phase and amplitude, instead of the
Wavelet method employed in 'dsCalcComodulgrams.m'.

Also, note that the function signature for 'mi_measure.m' has changed, since it
is supposed to use the sampling frequency 'Fs', but was not being passed it
before, for some unknown reason.